+++ R/RTL-package.R
@@ -5,5 +5,7 @@
 #' @import rlang
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
+#' @importFrom Rcpp sourceCpp
+#' @useDynLib RTL, .registration = TRUE
 ## usethis namespace: end
 NULL