+++ data-raw/DATASET.R
@@ -28,7 +28,8 @@
 # usethis::use_package("RCurl")
 # usethis::use_package("purrr")
 # usethis::use_package("zoo")
-#usethis::use_package("Rcpp")
+# usethis::use_package("Rcpp")
+# usethis::use_rcpp()
 
 # Setup RTL Webpage
 # usethis::use_pkgdown()